requirements.txt
langchain
langchain-community
faiss-cpu
sentence-transformers
google-generativeai
langgraph

